pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Minimata <minimata94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.12"}
uvicorn = "^0.34.0"
psycopg2 = "^2.9.10"
peewee = "^3.17.9"
python-dotenv = "^1.1.0"
httpx = "^0.28.1"
pytest = "^8.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
